fix: arch command and setup

+++ scripts/system/arch.bash
@@ -6,22 +6,31 @@ function system_arch_base_system() {
 	local -ra net=("openssh" "aria2" "curl" "wget")
 	local -ra files=("zip" "unzip" "bzip2" "tar")
 
-	pacman --noconfirm -Syyu && pacman --noconfirm -Sy "${security[@]}" "${build[@]}" "${net[@]}" "${files[@]}"
+	(sudo pacman --needed --noconfirm -Syu &&
+		sudo pacman --needed --noconfirm -Sy "${security[@]}" "${build[@]}" "${net[@]}" "${files[@]}") || exit ?
 }
 
 function system_arch_aur_helper_aura() {
 	local -r git_url="https://aur.archlinux.org/aura.git"
 	local -r clone_target_dir="/tmp/aura"
 
+	rm --verbose --recursive --force "${clone_target_dir}"
+	sudo pacman -Syu
+	if pacman -Q aura; then
+		return 0
+	fi
+
 	(git clone "$git_url" "$clone_target_dir" &&
 		cd "$clone_target_dir" &&
-		makepkg --syncdeps --noconfirm --install &&
+		makepkg --needed --syncdeps --noconfirm --install &&
 		rm --verbose --recursive --force "$clone_target_dir") || exit $?
 }
 
 function system_arch_qol() {
-	local -ra packages=("bash-completion" "man-db")
-	pacman --noconfirm -Syyu && pacman --noconfirm -Sy "${packages[@]}"
+	local -ra packages=("bash-completion" "man" "man-pages" "man-db")
+	(sudo pacman --needed --noconfirm -Syu &&
+		sudo pacman --needed --noconfirm -Sy "${packages[@]}" &&
+		sudo mandb -c) || exit $?
 }
 
 function system_arch_desktop_environment() {
@@ -29,7 +38,8 @@ function system_arch_desktop_environment() {
 	case "$desktop_environment" in
 	xfce | xfce4)
 		local -ra packages=("xfce4-goodies")
-		pacman --noconfirm -Syyu && pacman --noconfirm -Sy "${packages[@]}"
+		(sudo pacman --needed --noconfirm -Syu &&
+			sudo pacman --needed --noconfirm -Sy "${packages[@]}") || exit $?
 		;;
 	gnome) ;;
 	kde) ;;
